Fix a bug with references
Indexing modByte with bracket notation (e.g., modByte[0]) returns a std::bitset<N>::reference object
that always sets itself to the bit at the position it references. This means that the variable storing
the reference (in this case b0) is modified after a bit shift, thus returning the incorrect output. To
fix this, std::bitset<N>.test() is used to get the boolean state of the bit, which is static.